move convert member functions to util
